+++ R/plot.R
@@ -6,40 +6,57 @@
 #' @param x An object which is related to a geometric network
 #' (object of class \code{gn}, \code{gnpp} or \code{gnppfit}).
 #' @param ... Other arguments.
-#' @param covariate Character vector of length 1, name of the covariate
-#' that should be plotted.
-#' @param title Main title of the plot.
-#' @param title_x x-axis title (ignored if \code{frame = FALSE}).
-#' @param title_y y-axis title (ignored if \code{frame = FALSE}).
-#' @param size_lines Plotting size of the line segments. This specifies the
-#' \code{size} argument of \code{\link[ggplot2]{geom_segment}}.
-#' @param size_points Plotting size of the point pattern. This specifies the
-#' \code{size} argument of \code{\link[ggplot2]{geom_point}}.
-#' @param frame Should a frame be drawn around the network?
+#' @param title A named list with names "x", "y" and "plot" which specify the
+#' arguments \code{x}, \code{y} and \code{title} of the
+#' \code{\link[ggplot2]{labs}} function. Each list entry must
+#' be a character vector which has length equal to the number of plots.
+#' The list entries can remain unspecified in which case the respective titles
+#' are left blank.
+#' @param size A named list with names "lines" and "points" which specify the
+#' \code{size} argument of \code{\link[ggplot2]{geom_segment}} and
+#' \code{\link[ggplot2]{geom_point}}, respectively. Each list entry must be
+#' a numeric vector of length one with positive values. If the whole list or one
+#' entry remains unspecified, default values are used.
+#' @param color A named list with names "lines" and "points" which specify the
+#' \code{color} argument of \code{\link[ggplot2]{geom_segment}} and
+#' \code{\link[ggplot2]{geom_point}}. Each list entry must specify a valid color.
+#' By default, lines and points are plotted in black.
+#' @param shape The shape used for plotting the points. An integer between 0 and
+#' 25. Default to \code{shape = 1} which shows the points as a circle.
+#' @param frame If set to \code{TRUE}, draws a frame around the network and adds
+#' tick marks and axis labeling.
+#' @param covariate Character vector of length one which is name of the
+#' covariate to be plotted. Must be an external categorical covariate with at
+#' most ten different values.
+#' @param trans The transformation applied to the color bar of the intensity
+#' fit. Specifies the \code{trans} argument of
+#' \code{\link[ggplot2]{scale_color_gradient}}.
 #' @param sol Solution of the color network plot.
-#' @param scale The scale on which smooth terms should be plotted, either on
-#' the log scale (\code{scale = log}, default) or on the exp-scale
-#' (\code{scale = exp}).
 #' @param select Allows the plot for a single model term to be selected for
-#' printing. e.g. if you just want the plot for the second smooth term set
-#' \code{select = 2}.
+#' printing. e.g. if just the plot for the second smooth should be printed to
+#' the console, set \code{select = 2}.
 #' @return Invisibly returns an object of class \code{ggplot} or a list
 #' of \code{ggplot} objects.
 #' @import ggplot2
 #' @export
 
-plot.gn <- function(x, ..., title = "", title_x = "", title_y = "",
-                    size_lines = 1, frame = TRUE) {
+plot.gn <- function(x, ..., title = list(), size = list(), color = list(),
+                    frame = FALSE) {
   v1_x <- v1_y <- v2_x <- v2_y <- NULL
   g <- ggplot(x$lins)
   if (!frame) {
     g <- g + theme_void()
   } else {
     g <- g + theme_bw()
+    if (is.null(title$x)) title$x <- "x"
+    if (is.null(title$y)) title$y <- "y"
   }
-  g <- g + geom_segment(aes(x = v1_x, y = v1_y, xend = v2_x, yend = v2_y), size = size_lines,
+  if (is.null(size$lines)) size$lines <- 1
+  if (is.null(color$lines)) color$lines <- "black"
+  g <- g + geom_segment(aes(x = v1_x, y = v1_y, xend = v2_x, yend = v2_y),
+                        size = size$lines, color = color$lines,
                         lineend = "round", linejoin = "bevel") +
-    labs(x = title_x, y = title_y, title = title) +
+    labs(x = title$x, y = title$y, title = title$plot) +
     theme(panel.grid = element_blank(),
           plot.title = element_text(hjust = 0.5))
   print(g)
@@ -50,9 +67,8 @@ plot.gn <- function(x, ..., title = "", title_x = "", title_y = "",
 #' @import ggplot2
 #' @export
 
-plot.gnpp <- function(x, ..., covariate = NULL,
-                      title = "", title_x = "", title_y = "",
-                      size_lines = 1, size_points = 1, frame = TRUE) {
+plot.gnpp <- function(x, ..., title = list(), size = list(), color = list(),
+                      shape = 1, frame = FALSE, covariate = NULL) {
   if (length(covariate) > 1){
     stop("Currently, only one covariate can be plotted.")
   }
@@ -62,18 +78,33 @@ plot.gnpp <- function(x, ..., covariate = NULL,
     g <- g + theme_void()
   } else {
     g <- g + theme_bw()
+    if (is.null(title$x)) title$x <- "x"
+    if (is.null(title$y)) title$y <- "y"
   }
+  if (is.null(size$lines)) size$lines <- 1
+  if (is.null(size$points)) size$points <- 2.5
+  if (is.null(color$lines)) color$lines <- "black"
+  if (is.null(color$points)) color$points <- "black"
   g <- g + geom_segment(aes(x = v1_x, y = v1_y, xend = v2_x, yend = v2_y),
-                          size = size_lines, lineend = "round", linejoin = "mitre") +
-    labs(x = title_x, y = title_y, title = title) +
+                        size = size$lines, lineend = "round", linejoin = "mitre") +
+    labs(x = title$x, y = title$y, title = title$plot) +
     theme(panel.grid = element_blank(),
-                   plot.title = element_text(hjust = 0.5))
+          plot.title = element_text(hjust = 0.5))
   if (is.null(covariate)) {
-    g <- g + geom_point(data = x$data, aes(x = x, y = y),
-                        color = "red", size = size_points)
+    g <- g + geom_point(data = x$data, aes(x = x, y = y), shape = shape,
+                        color = color$points, size = size$points)
   } else {
-    g <- g + geom_point(data = x$data, aes(x = x, y = y, color = !!sym(covariate)),
-                        size = size_points)
+    if (is.factor(X$data[[covariate]]) | is.character(X$data[[covariate]])) {
+      k <- length(unique(X$data[[covariate]]))
+      if (k > 10) stop("Number of different values must be at most ten.")
+      g <- g + geom_point(data = x$data, aes(x = x, y = y,
+                                             color = !!sym(covariate),
+                                             shape = !!sym(covariate)),
+                          size = size$points) +
+        scale_shape_manual(values = 0:(k-1))
+    } else {
+      stop("Currently only categorical covariates are supported.")
+    }
   }
   print(g)
   invisible(g)
@@ -85,8 +116,9 @@ plot.gnpp <- function(x, ..., covariate = NULL,
 #' @importFrom splines splineDesign
 #' @export
 
-plot.gnppfit <- function(x, ..., select = NULL, title = "", title_x = "x", title_y = "y",
-                         size_lines = 1, frame = TRUE, sol = 100, scale = "log") {
+plot.gnppfit <- function(x, ..., title = list(), size = list(), color = list(),
+                         shape = 1, frame = FALSE, data = FALSE,
+                         trans = "identity", select = NULL,  sol = 100) {
 
   stopifnot(inherits(x, "gnppfit"))
   e <- xend <- y <- yend <- intensity <- lower <- upper <- NULL
@@ -104,9 +136,9 @@ plot.gnppfit <- function(x, ..., select = NULL, title = "", title_x = "x", title
       yy <- lins_m$v1_y[i] + tt*dy[i]
       zz <- cs[i] + (tt - 1/(2*sol))[-1]*lins_m$length[i]
       df[[1]] <- bind_rows(df[[1]], tibble(id = lins_m$id[i], e = m,
-                                               x = utils::head(xx, -1), xend = xx[-1],
-                                               y = utils::head(yy, -1), yend = yy[-1],
-                                               z = zz))
+                                           x = utils::head(xx, -1), xend = xx[-1],
+                                           y = utils::head(yy, -1), yend = yy[-1],
+                                           z = zz))
     }
   }
 
@@ -121,13 +153,23 @@ plot.gnppfit <- function(x, ..., select = NULL, title = "", title_x = "x", title
     g[[1]] <- g[[1]] + theme_void()
   } else {
     g[[1]] <- g[[1]] + theme_bw()
+    if (is.null(title$x)) title$x <- "x"
+    if (is.null(title$y)) title$y <- "y"
   }
+  if (is.null(size$lines)) size$lines <- 1
   g[[1]] <- g[[1]] + geom_segment(aes(x = x, xend = xend, y = y, yend = yend, color = intensity),
-                          size = size_lines, lineend = "round", linejoin = "mitre") +
-    labs(color = "Intensity", x = title_x, y = title_y, title = title) +
+                                  size = size$lines, lineend = "round", linejoin = "mitre") +
+    labs(color = "Intensity", x = title$x[1], y = title$y[1], title = title$plot[1]) +
     theme(panel.grid = element_blank(),
-                 plot.title = element_text(hjust = 0.5)) +
-    scale_color_gradient(low = "blue", high = "red")
+          plot.title = element_text(hjust = 0.5)) +
+    scale_color_gradient(low = "blue", high = "red", trans = trans)
+  if (data) {
+    if (is.null(size$points)) size$points <- 2.5
+    if (is.null(color$points)) color$points <- "black"
+    g[[1]] <- g[[1]] + geom_point(data = X$data, aes(x = x, y = y),
+                                  size = size$points, color = color$points,
+                                  shape = shape)
+  }
 
   if (length(g) > 1){
     for (i in 2:length(g)) {
@@ -141,13 +183,11 @@ plot.gnppfit <- function(x, ..., select = NULL, title = "", title_x = "x", title
       y <- as.vector(X%*%theta)
       limits <- confidence_band(theta, V, X, R = 1000)
       df[[i]] <- tibble(x = xx, y = y, lower = limits$lower, upper = limits$upper)
-      if (scale == "exp"){
-        df[[i]] <- mutate(df[[i]], y = exp(y),
-                          lower = exp(lower), upper = exp(upper))
-      }
+
       g[[i]] <- ggplot(df[[i]]) +
         geom_ribbon(aes(x = x, ymin = lower, ymax = upper), fill = "grey50") +
         geom_line(aes(x = x, y = y), color = "red") +
+        labs(x = title$x[i], y = title$y[i], title = title$plot[i]) +
         theme_bw()
     }
   }
@@ -160,6 +200,7 @@ plot.gnppfit <- function(x, ..., select = NULL, title = "", title_x = "x", title
       ask <- FALSE
     }
   }
+  if (length(g) == 1) g <- g[[1]]
   invisible(g)
 }
 
@@ -167,8 +208,9 @@ plot.gnppfit <- function(x, ..., select = NULL, title = "", title_x = "x", title
 #' @import ggplot2
 #' @export
 
-plot.lppfit <- function(x, ..., title = "", title_x = "x", title_y = "y",
-                        size_lines = 1, frame = TRUE, sol = 100, scale = "log") {
+plot.lppfit <- function(x, ..., title = list(), size = list(), color = list(),
+                        shape = 1, frame = FALSE, data = FALSE,
+                        trans = "identity", sol = 100) {
   e <- xend <- y <- yend <- intensity <- lower <- upper <- NULL
   G <- as_gn(x)
   f <- as.linfun(x)
@@ -185,18 +227,33 @@ plot.lppfit <- function(x, ..., title = "", title_x = "x", title_y = "y",
       yy <- lins_m$v1_y[i] + tt*dy[i]
       zz <- cs[i] + (tt - 1/(2*sol))[-1]*lins_m$length[i]
       df <- bind_rows(df, tibble(id = lins_m$id[i], e = m,
-                                      x = utils::head(xx, -1), xend = xx[-1],
-                                      y = utils::head(yy, -1), yend = yy[-1],
-                                      z = zz))
+                                 x = utils::head(xx, -1), xend = xx[-1],
+                                 y = utils::head(yy, -1), yend = yy[-1],
+                                 z = zz))
     }
   }
   df$intensity <- f(df$x, df$y)
-  g <- ggplot(df) + geom_segment(aes(x = x, xend = xend, y = y, yend = yend, color = intensity),
-                                  size = size_lines, lineend = "round", linejoin = "mitre") +
-    labs(color = "Intensity", x = title_x, y = title_y, title = title) +
-    theme_void() +
+  if (!frame) {
+    g <- ggplot(df) + theme_void()
+  } else {
+    g <- ggplot(df) + theme_bw()
+    if (is.null(title$x)) title$x <- "x"
+    if (is.null(title$y)) title$y <- "y"
+  }
+  if (is.null(size$lines)) size$lines <- 1
+  g <- g + geom_segment(aes(x = x, xend = xend, y = y, yend = yend, color = intensity),
+                    size = size$lines, lineend = "round", linejoin = "mitre") +
+    labs(color = "Intensity", x = title$x, y = title$y, title = title$title) +
     theme(panel.grid = element_blank(),
           plot.title = element_text(hjust = 0.5)) +
-    scale_color_gradient(low = "blue", high = "red")
+    scale_color_gradient(low = "blue", high = "red", trans = trans)
+  if (data) {
+    if (is.null(size$points)) size$points <- 2.5
+    if (is.null(color$points)) color$points <- "black"
+    g <- g + geom_point(data = X$data, aes(x = x, y = y),
+                                  size = size$points, color = color$points,
+                                  shape = shape)
+  }
+  print(g)
   invisible(g)
 }